+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+* Fixed tracing of project namespaces.
+* Added truncation of var names so they display better.
+
 ## 0.1.0 - First Release
 * Every feature added
 * Every bug fixed